minor fixes and separate into cmd init metadata
