+++ .gitignore
@@ -18,4 +18,4 @@
 
 go.sum
 
-.env
+cmd/.env
