+++ main.go
@@ -1,15 +1,13 @@
 package main
 
 import (
-	"flag"
 	"fmt"
-	"github.com/joho/godotenv"
-	v1 "ios-backend/src/CoinBaseApiRequests/v1"
 	"net/http"
 	"os"
 
 	"github.com/gorilla/mux"
 	"github.com/jmoiron/sqlx"
+	_ "github.com/lib/pq"
 	"github.com/rs/zerolog/log"
 	"ios-backend/src/configs"
 	"ios-backend/src/user/delivery"
@@ -20,12 +18,6 @@ import (
 	currRepo "ios-backend/src/currency_info/repository"
 	currUCase "ios-backend/src/currency_info/usecase"
 )
-
-func init() {
-	if err := godotenv.Load(); err != nil {
-		log.Print("No .env file found")
-	}
-}
 
 func main() {
 	r := mux.NewRouter()
@@ -51,14 +43,6 @@ func main() {
 		return
 	}
 
-	init := flag.Bool("metadata_init", false, "Insert 200 cryptocurrencies metadata info")
-	flag.Parse()
-	if *init {
-		err = v1.GetCurrencyMetadata(conn)
-		if err != nil {
-			log.Error().Msgf(err.Error())
-		}
-	}
 
 	rep := repository.NewPostgresUserRepository(conn)
 	ucase := usecase.NewUserUsecase(rep, timeoutContext)
