+++ src/CoinBaseApiRequests/v1/get_crypto_metadata.go
